Modify the functionality and name of a method for another much more appropiated.
Actually, the disconnectAgent() method must be named disconnect()
due to this method is invoked when an HDP channel is being released.
In other words, an HDP channel is being disconnected at this point
but the agent could be connected to the manager through others HDP
channels, so, we mustn't delete the agent from our list. In fact,
we have an HDP agent when we are paired with a device which exposes
a Bluetooth Health Profile, and it is not necessary to stablish an
HDP channel to add it to the list of managed agents.